+++ src/pyworkflowkit/adapters/metadata/__init__.py
@@ -5,10 +5,13 @@ from pyworkflowkit.adapters.metadata.sqlalchemy import (
     SqlAlchemyMetadataStore,
     SqlAlchemyUnitOfWork,
 )
+from pyworkflowkit.adapters.metadata.sqlite import SQLiteMetadataStore, SQLiteSettings
 
 __all__ = [
     "MemoryMetadataStore",
     "MemoryUnitOfWork",
+    "SQLiteMetadataStore",
+    "SQLiteSettings",
     "SqlAlchemyMetadataStore",
     "SqlAlchemyUnitOfWork",
 ]